Crate pinned_init

This library is used to facilitate safe pinned initialization for as many parties involved as possible. The process of initialization follows the following steps:

You create a FooUninit<...> (an alias for Foo<..., false>)

You move the FooUninit<...> behind a unique owning pointer (here named SomePtr<FooUninit<...>>)

You use the init() function of SomePtr<FooUninit<...>> provided by the SafeInit trait, to initialize the contents 3a. This library creates a NeedsPinnedInit<FooUninit<...>> from the SomePtr<FooUninit<...>>. 3b. This library uses the NeedsPinnedInit<FooUninit<...>> to call the init_raw function of FooUninit<...> which initializes it. 3c. This library transmutes the SomePtr<FooUninit<...>> to SomePtr<Foo<...>> and returns that.

This library aims to make it ergonomic and safe to not only write code using pinned initialized types, but also to create them. When you embed a pinned init type within another struct, you can use the #[pinned_init] macro to turn the outer struct into a pin initialized type and automatically initializes the inner type.

When you have more compilcated requirements (for example you need a self referential struct) you probably cannot avoit writing unsafe code, but the pinned initialization part will probably not require any unsafe if you follow these guidelines:

use #[manual_init] to add the minimal comapability for this libarary, this will:

put #[pin_project] your struct, you will have to manually add #[pin] whenever you desire structual pinning

derive some traits from this lib: transmute and manualinithelper

implement PinnedInit for your type and use the begin_init() function of NeedsPinnedInit to access the fields you need for initialization, when you need access to other fields of the type, then TODO

Modules

needs_init

Custom pointer types used to ensure that initialization was done completly. The pointer types provided by this module [NeedsPinnedInit<'init, T>] and [NeedsInit<'init, T>] will

ptr

Module providing a special pointer trait used to transfer owned data and to allow safe transmuation of data without forgetting to change other pointer types.

static_uninit

Using MaybeUninit<T> requires unsafe, but this is often not necessary. This module provides StaticUninit<T, INIT> a safer alternative using static type checking to ensure one cannot use an uninitialized value as an initialized and to prevent leaking values when initializing a value twice without dropping the contents.

transmute

Initializing a value in place (because it is pinned) requires a transmuation. Because transmuations are inherently unsafe, this module aims to create a safer abstraction and requires users to explicitly opt in to use this initialization.

Traits

PinnedInit

Facilitates pinned initialization. Before you implement this trait manually, look at the pinned_init proc macro attribute, it can be used to implement this trait in a safe and sound fashion in many cases.

SafePinnedInit

Sealed trait to facilitate safe initialization of the types supported by this crate.

Attribute Macros

manual_init

Use this attribute on a struct with named fields to ensure safer pinned initialization of all the fields marked with #[init].

pinned_init

Use this attribute on a struct with named fields to ensure safe pinned initialization of all the fields marked with #[init].