Trait pinned_init::ptr::OwnedUniquePtr

pub unsafe trait OwnedUniquePtr<T: ?Sized>: DerefMut<Target = T> + Sized where
    Self: TypesEq<Self::Ptr<T>>, {
    type Ptr<U: ?Sized>: DerefMut<Target = U>;
}

A (smart) unique pointer which owns its data (e.g. [alloc::boxed::Box]). This pointer provides access to T via DerefMut.

Safety

All types implementing this trait need to

own the data they point to.

be the only way to access the data behind this pointer.

provide the same pointer type as Self with only a different pointee via the Self::Ptr associated type.

Associated Types

type Ptr<U: ?Sized>: DerefMut<Target = U>

Access the same underlying pointer type with a different pointee type. Self == Self::Ptr<T>