Trait pinned_init::transmute::TransmuteInto

pub unsafe trait TransmuteInto<T>: Sized {
    unsafe fn transmute_ptr(this: *const Self) -> *const T;

    unsafe fn transmute(self) -> T { ... }
}

Marks and allows easier unsafe transmutation between types. When implementing this type manually, which should only be done for custom pointer types which also implement crate::ptr::OwnedUniquePtr. When you use the proc macro attributes [pinned_init] and [manual_init] this trait will be implemented automatically to transmute from the uninitialized to the initialized variant.

When using this trait, it is not required to use one of the provided functions, you may mem::transmute a value of Self to T, you may use a union to reinterpret a Self as a T and you may use pointer casts to cast a pointer to Self to a pointer to T. Of course you will need to still ensure the safety requirements of this trait.

Safety

When implementing this type the following conditions must be true:

T and Self have the same layout.

transmutation is only sound, if all invariants of T are satisfied by all values of Self transmuted with this trait.

Required methods

unsafe fn transmute_ptr(this: *const Self) -> *const T

Unsafely transmutes a pointer to Self to a pointer to T.

Safety

T and Self must have the same layout.

All invariants of T need to be satisfied by the value at this.

Provided methods

unsafe fn transmute(self) -> T

Unsafely transmutes self to T

Safety

T and Self must have the same layout.

All invariants of T need to be satisfied by self.